requirements.txt
absl-py==2.2.2
accelerate==0.24.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
datasets==2.14.6
dill==0.3.7
filelock==3.13.1
frozenlist==1.6.0
fsspec==2023.10.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-auth-oauthlib==1.0.0
grpcio==1.71.0
huggingface-hub==0.17.3
idna==3.4
importlib_metadata==8.6.1
Jinja2==3.1.4
joblib==1.4.2
Markdown==3.8
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.15
networkx==3.2.1
numpy==1.23.5
oauthlib==3.2.2
packaging==25.0
pandas==1.5.3
pillow==11.0.0
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.1
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.13.1
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.14.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.65.0
transformers==4.34.1
typing_extensions==4.12.2
urllib3==1.26.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
